[FedCM] Modify the loginHint implementation
This CL changes the implementation of loginHint so that it requires that
the accounts response provide the valid hints for accounts, and the JS
call only requires passing the loginHint. Since isRequired does not seem
to be required, it is removed, with the behavior being set as if it had
been set to true.

Bug: 1356021
Change-Id: Ie5cc5f5be53e55b838ec67c4cfe1ba0d978f4218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4447539
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1134805}
NOKEYCHECK=True
GitOrigin-RevId: 0d491060336a6a034b60b8425d0a173460d0a289