blink: make Element::BeginParsingChildren() concrete
There was only one override, and I believe it can be switched
to ParserDidSetAttributes. The order is:

1. ParserDidSetAttributes().
2. call AttributeChanged() for all attributes.
3. BeginParsingChildren().

This patch moves processing of directionality from 3 to 1.

The patch that made BeginParsingChildren() virtual is
https://chromium-review.googlesource.com/c/chromium/src/+/2666349 .

pinpoint does not really show much of a gain here from what was
lost by calling Begin/FinishParsingChildren(), but this is better
than nothing.

Bug: 1422666

Change-Id: I8b7559c39c1593c3046c22cf426164b035cd28e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4469869
Reviewed-by: Mason Freed <masonf@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135287}
NOKEYCHECK=True
GitOrigin-RevId: 24d6440de090bba70dda37960635725d45dd58cd