[@container] Use correct style container for ::part()/::slotted()
We always used ParentOrShadowHost() to find the first candidate for
style containers, but that is not correct for ::part() or ::slotted()
rules which should start at the originating element for the selector.

Also, make sure we do not overwrite other parts of StyleRecalcContext
when updating StyleRecalcContext for matching shadow pseudo element
rules.

Bug: 1302630
Change-Id: I6fd3a0d51df12cc6dbc7cb9c1e88ecccac08cc44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4459274
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1135311}
NOKEYCHECK=True
GitOrigin-RevId: e004312a88e3276d0e9dd0500cf9f3168c08bbb3