[RemoveLegacy] Fix crash at subtree root without layout-result.
It was possible (with fieldsets) to layout at a subtree root which
didn't undergo layout previously (due to the way children are created).

Check that we have a previous layout result before selecting the subtree
root.

This is due to the new LayoutNewSubtreeRoot logic, and hence added it
under that flag.

Fixed: 1439692
Change-Id: I54eab45b213616f0ad174fbdf0d6cdaf8916c8e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4472060
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1135370}
NOKEYCHECK=True
GitOrigin-RevId: 5b52c45f71b77f9c7c96835a296480d02821a179