Improve the loop detection for popovers
See [1] for much more context, but the prior CL [2] did not completely
solve the problem. This CL adds a bit to PopoverData that keeps track
of whether we're currently showing or hiding a popover. If that bit
is already true on a subsequent call, it means we have a loop, and in
that case events are disabled and a console warning is generated. In
addition, the loop added in [2] is moved into HideAllPopoversUntil(),
so that it functions for both show and hide.

The test is enhanced to make sure the behavior is interoperable.

[1] https://github.com/whatwg/html/issues/9197
[2] https://chromium-review.googlesource.com/c/chromium/src/+/4472129

Bug: 1307772
Change-Id: I7c9efb8eea3e5dea04c7998e200f0e9e458ef966
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4471675
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135651}
NOKEYCHECK=True
GitOrigin-RevId: 6753b11ee7b8e4e24ef12b52eb0bf2c2c24b96fe