[css-flex] Disable fixed-table's infinite size for intrinsic flex sizing
A fixed-table's intrinsic max-content is ~infinite when it has a
%-width. That's been around for a while and ships in all browsers.

But it is problematic for the new flex intrinsic sizing algorithm
because now the infinity percolates up to the flex container's
min-content size[1]. We obviously don't want a min-content size of
infinity and gmail relies on the old behavior.

So this is a hacky patch that makes fixed-tables max-content size be the
same as auto-tables when an ancestor flex table is calculating its
intrinsic sizes. It will probably be web compatible, but unsure.

This affects both the min and max intrinsic size of the flex container.

[1] New flex min-content algorithm accounts for item's flex base sizes.
A %-width fixed-table's flex base size is its max-content size, which,
as mentioned above, is ~infinity.

Change-Id: I7010fa73df8e88a4812b23780e5d58a3f63da730
Bug: 240765
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4420422
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135691}
NOKEYCHECK=True
GitOrigin-RevId: 008c0e9740d071aed121ac1a01b404114d776a96