Add new flag for tabbed mode customizations.
Add a new flag for tabbed mode to separate the 'tabbed' display mode
and the 'tab_strip' manifest field. This new flag is for the
'tab_strip' field which allows apps to customise the tab strip and the
existing flag, kDesktopPWAsTabStrip, is for the tabbed display mode.

See explainer for more detail:
https://github.com/WICG/manifest-incubations/blob/gh-pages/tabbed-mode-explainer.md

Bug: 897314
Change-Id: I86806d626ba2357fdcbf01f1b29512aa6f09015d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4478391
Commit-Queue: Louise Brett <loubrett@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135781}
NOKEYCHECK=True
GitOrigin-RevId: 0115847959389e0596cc2cea975f1bd7728ceaa5