Move cursor constraints from WebCursor to ui::Cursor
When creating a custom cursor, `content::WebCursor` places some limits
on its parameters. Namely, a maximum size in DIPs for the cursor image
and a lower bound on the scale factor so that the previous value
doesn't overflow. Additionally, the hotspot is always clamped to the
image dimensions.

In order to advance the removal of WebCursor, this CL moves these checks
to `ui::Cursor`. The hotspot will be clamped when building a custom
cursor, and the image size will be checked in a new static method
`AreDimensionsValidForWeb`, which will be used both by Blink and the
cursor mojom traits, so that even compromised renderers cannot bypass
these limits.

Bug: 1149906
Change-Id: Id4e15aec574d26041f3f340d71ac7627ef361896
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4241809
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136433}
NOKEYCHECK=True
GitOrigin-RevId: e77d8705ea2d252a74e0fbd7778e653c010f5c53