Replace test-only AttributionManager-reset with WebTestControlHost reset
Like cookies, persistent Attribution Reporting data needs to be
cleared between web tests in order to avoid interference. We currently
achieve this with a test-only JS method, but it is simpler to do so
in the same way that cookies are cleared in the WebTestControlHost.

This also has the benefit of reducing the API surface of the
StoragePartition, AttributionManager, and AttributionStorageSql
interfaces while simultaneously using only the public APIs of each of
them.

For tests that do not use Attribution Reporting (which is the vast
majority of them), the runtime cost of this is effectively the cost of
posting a no-op task to the Attribution Reporting storage sequence.

Bug: 1439886
Change-Id: I8cb5d3bcc99b6a77a2e9731297624230c4b1645e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4476424
Quick-Run: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136772}
NOKEYCHECK=True
GitOrigin-RevId: 0c7a491b423660fe5d083d005d67372910321395