media: Rename kIsVideoEncrypted to kIsCdmAttached
This is reported when a CDM is attached, which doesn't necessarily mean
that the video is encrypted. It could happen for a clear stream, or an
audio-only stream.

This CL renames it to match it's meaning.

Bug: 1435070
Change-Id: I77cc3276ef08eb5c1b5513b3cd29347cd763b414
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455712
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136875}
NOKEYCHECK=True
GitOrigin-RevId: 703b760c6369b00ac058da04316dda827624d4fd