Remove Ukm source id type SOFT_NAVIGATION_ID
This CL essentially is a revert of this previous CL that add a new UKM
source id type SOFT_NAVIGATION_ID.
See https://chromium-review.googlesource.com/c/chromium/src/+/4427499.

The type was added for the UpdateSourceUrl method to be used for soft
navigation metric use cases. But Sun pointed out an alternative way. As
a result this type is no longer needed.

Since the there are date and version update made by running a script in
both previous CL and this CL. We can’t do a revert.

Bug: 4381537
Change-Id: I18cc9dccc039ae9bcbec0573b2977fd7dc1e6634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4482605
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1137171}
NOKEYCHECK=True
GitOrigin-RevId: 96a1d557a99f36a1b2e0fa43d472f35c74837f9a