[FedCM] Allow closing of modal dialogs using web contents
We want to open a modal dialog for users to sign in and we want to be
able to close the same modal dialog. After the user has signed in, we
would use web contents to retrieve the modal dialog and prompt the
delegate to close the dialog.

Bug: 1430830
Change-Id: Ic6311c3f78129e3e2cfb99a9ef8eacae4d9bf10b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4414977
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Zachary Tan <tanzachary@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1137427}
NOKEYCHECK=True
GitOrigin-RevId: 5fcb2bffb2ae6da7506248ae832538aac2b890f4