DevTools: Persist devtools_navigation_token in RFH
We currently use devtools_navigation_token to associate a preloading
attempt with its initiator document. However, we currently send the
token from the renderer back to the browser through a mojo IPC
(SpeculationHost.UpdateSpeculationCandidates), despite this token
having been generated in the browser in the first place (in most
scenarios). With this change, we can now retrieve the token directly
from the initiating RFH in the browser process, instead of resending it
back from the renderer.

This CL stores the devtools_navigation_token from a NavigationRequest
in a RFH after a navigation commits. The token has "per-document"
semantics and is stored in RFH::DocumentAssociatedData. The token isn't set for RFHs with the initial empty document (in these scenarios, the
token is generated in the renderer, and the browser doesn't currently
have access to it). We currently don't have a use case that needs to
access the token in such scenarios, but we could revisit this if we
need to in the future.

Bug: 1384419
Change-Id: I880d295473cf9333795ceab947f1089be366ecb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4456927
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1137876}
NOKEYCHECK=True
GitOrigin-RevId: 94ec2fea918048f859e1ddd784a75bc250a8bb78