Rename Trigger Attestation to Report Verification
Report Verification now refers to the whole feature. Trigger
verification is also used to refer to the first portion of the operation
which verifies a trigger that is then embedded with the report.

Feature name, histogram names & proto field were also renamed.

The feature is not used yet.

Bug: 1425187
Change-Id: I690435b42fd8a57b2c5706ccfd1bf7125b94b737
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4469713
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1137913}
NOKEYCHECK=True
GitOrigin-RevId: 6e68d28de3871164e4a6c779a977ca50c5d453a0