Don't bounds-check a history API navigation's offset in the renderer
When a history API navigation occurs (e.g., history.go(-2)), we
currently check whether the given offset is in-bounds before sending
the navigation to the browser. The spec and WPT expect an async step
to occur (i.e., going to the browser) so that we can take into
account any pending back/forward navigations in offset calculation.

However, when a history API navigation begins, we pause virtual time
if the navigation is in-bounds only. If we can't check that in
the renderer, we need of notifying the renderer that a history API
navigation was out of bounds so that we can unpause virtual time.
Add a case to the TraverseCancelled() message to unpause virtual
time and call it for out-of-bounds history API navigations.
(Note that TraverseCancelled() is currently only used for navigation
API traversals, but there's nothing stopping us from using the same
plumbing for this case).

This change gives us the correct behavior when history.back()
and history.forward() are called in rapid succession while at the
front of the back/forward list. The forward() should be calculated
from the pending entry for the back() navigation, resulting in
the two navigations cancelling each other out. The renderer doesn't
know what the pending offset is, so it thinks the forward() is trying
to go out-of-bounds and incorrectly rejects that navigation.

Change-Id: I36bfbe8211df5cad9f6cefeee659514299805434
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496257
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138110}
NOKEYCHECK=True
GitOrigin-RevId: 758e5e482e7cbbfd64073768da3f24c2788ffa42