[FSA] Update FileSystemWritableFileStream.write() to set the correct offset value
Update the offset value to the correct position for write operations.
Currently, the offset is not updated which results in behavior that
does not align with the specifications for FileSystemWritableFileStream.

Add tests to ensure that multiple write operations within the same
writable execute as expected.

Bug: 1427819
Change-Id: I6f6913e2ac039a796e3bac7455fedb92f8439eb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4409002
Commit-Queue: Christine Smith <christinesm@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138324}
NOKEYCHECK=True
GitOrigin-RevId: d41981cb1db6bffa93fb0cb4223257d1378c943a