Do not crash if image first visibility detected twice
Lazy load image visibility information is tracked with an intersection
observer on the local root (visibility_metrics_observer_), and the data
is stored on the Element itself (see: visible_load_time_metrics_ on
HTMLImageElement). This visibility information is only used for metrics.
This patch fixes a crash when the visibility information was set twice
due to two visibility metrics intersection observers.

How can there be two visibility metrics intersection observers? An
element's local root can change if the element moves between frames,
such as if an element is moved from a2 to a1 in the following tree:
  Frame a1 (local root)
    Frame b (remote)
      Frame a2 (local root)
An element can be moved between a1 and a2, causing one visibility
metrics intersection observer for each local root.

A regression test was used to write this patch (crrev.com/c/4498925)
but the test is complex and brittle. The maintenance cost of this test
is not worth the coverage for this rare metrics DCHECK.

Bug: 1440980
Change-Id: I119f6109b247874224df465f7f2d6134d4d9e9c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4497540
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138444}
NOKEYCHECK=True
GitOrigin-RevId: 46e5c4103744897318bb9fe77877a36a98a71a1a