[Camera] Deprecate VideoFrameFeedback::mapped_sizes.
The feature that added mapped_sizes (InCapturerScaling) was not shipped,
so this field can be deprecated.   The code that consumed mapped_sizes
in the MF capturer was removed in crrev.com/r/4482684, so we can also
remove the code in Blink that sets them.

Note we cannot actually remove the field from the mojo struct as it's
part of a stable message.

Bug: 1440076
Change-Id: I4402b4b168d8b5f6683cb1ddea413fbc2b423fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4491265
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138616}
NOKEYCHECK=True
GitOrigin-RevId: 2a1656febee6b99fee58d15db83d69142d8c20d5