Support non-preferred canvas formats
This CL supports required formats that aren't supported by the
underlying OS (`bgra8unorm` on Android, `rgba8unorm` on Mac) by
allocating a temporary texture of the desired format and copying
it to a supported texture format when it's time to present.

This should be seen at a fallback, last resort path, and if
more efficient ways of supporting the formats in question are
available they should be preferred.

Also adds a mechanism for reporting singleton warnings: Messages
that only show once per device to avoid spamming the user but
can provide performance/correctness nudges.

Bug: 1361468
Bug: 1298618
Change-Id: I836a162115bf934b361e508d96974140a5ecfd07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4491309
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138751}
NOKEYCHECK=True
GitOrigin-RevId: 6130bccc1813f386308c3a5771e90f8271b19c5c