Let resolveToConfig take a promise.
This change modifies the resolveToConfig field in the auction config
struct to accept a promise as well as a boolean.

The move with fenced frames to replace the src attribute with
FencedFrameConfig isn't compatible with iframes (we are choosing to not
support FencedFrameConfigs with iframes). This creates a problem
if the caller needs to know what kind of frame is being loaded at the
time of calling runAdAuction, which is typically determined by a
contextual signal or experiment states. However, these signals are only
available after receiving the contextual response.

The solution is to let resolveToConfig be a Promise that resolves with
a boolean received after the contextual request. This change will enable
the caller to determine the rendering mode at runtime, allowing for the
winner ad to be rendered in either iframes or fenced frames. This
approach will also make the transition smoother for ad tech companies,
who can unconditionally pass {resolveToConfig: true} if
FencedFrameConfig is supported.

Change-Id: I03f6b904e28d15216c03f1295e857860fbdf9b6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4499266
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138991}
NOKEYCHECK=True
GitOrigin-RevId: 3f9eb6d700c819b6edfb5123f8e5597d3f65b9ba