[FedCM] Implement preventSilentAccess for FedCM
This patch implements preventSilentAccess for FedCM. When the API is
called (preferably upon user signing out of a site), auto re-authn will
be disabled until next time the user finishes a FedCM flow.

This is operated on the site level instead of origin level. e.g. if the
API is called on https://foo.rp.example, auto re-authn will be disabled
on https://rp.example.

Note that this is a separate control and does not conflict with the
10 mins quiet period or others.

Bug: 1439238
Change-Id: Ib5a9b6fb7aec0b324bb9e8f945511084d1f7db41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4493345
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139019}
NOKEYCHECK=True
GitOrigin-RevId: ab5515f7a00b3285b4b9bed0b6b6af1aede456b6