Correctly account for shared memory in ArrayBufferContents
When an ArrayBufferContents is created from shared memory, we use
PlatformSharedMemoryHandle::MapAt to map the shared memory into the
virtual address space, and this will also account for the new mapping in
the SharedMemorySecurityPolicy. However, when we later unmap the shared
memory mapping again, we were previously directly using the
SharedMemoryMapper, and so the unmapping was not accounted for in the
SharedMemorySecurityPolicy. This CL fixes that by introducing a new
PlatformSharedMemoryRegion::Unmap helper function that performs the
accounting and forwards to the provided SharedMemoryMapper to do the
actual unmapping of the shared memory pages.

Bug: chromium:1433138
Change-Id: I628636688224410d9e28928b467e8b914a589dea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4488907
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139083}
NOKEYCHECK=True
GitOrigin-RevId: 8d8f2b3093f19e588a35c5d9a91f91bdb713738d