[no-vary-search-hint] Hint default value should not fail parsing
At the moment, we are failing parsing if the author adds a
No-Vary-Search hint that is equivalent to the default URL search variance.

This CL removes the error message that appears in the console, and
makes sure the speculation candidate is successfully parsed and it is
populated with the default URL search variance when No-Vary-Search
hint is equivalent to the default.

Bug: 1441644
Change-Id: I3e2b3a244505806ff62ab488ede0712db79a0a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4500326
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139193}
NOKEYCHECK=True
GitOrigin-RevId: 795dbdb5eb24982d22b24445f390f8954183a14d