Warn if 'depth' key used in GPUExtend3DDict
If the user passes a key of 'depth' it's almost certainly a bug, since
the correct key is 'depthOrArrayLayers'. Instead of failing silently,
adding this key as an `any` to the IDL allows us to detect when it's
used and fire a one-time warning message indicating the likely error
to the developer.

Bug: 1440900
Change-Id: I22941cf8d8c4a2abbd7d9a12d2b38f400435eb1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4503159
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139234}
NOKEYCHECK=True
GitOrigin-RevId: 7568d7b0f4ff23fb67798a115bd3d765ffb70c6b