[fedcm] Expose the IdentityProvider.resolve() WebIDL JS entrypoint
The AuthZ API [1] allows IdPs to load their own content in a modal
dialog to ask for the user's permissions when necessary. So, instead of
returning an idtoken from the HTTP endpoint, they can return an idtoken
from the content area loaded in the modal dialog.

This CL exposes the JS API, and leaves the implementation left to
another CL (we are going to have to expose the mojom service and
implement it too -- so, multitple steps).

[1] https://docs.google.com/document/d/1uSHNIT2wZj0ojmelsOa698oZQ2XC7VLM6WnxYkHU4fk/edit

Bug: 1429083
Change-Id: Ice75dccea6221b684afae96e807268b769a1b63d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4482737
Commit-Queue: Sam Goto <goto@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139252}
NOKEYCHECK=True
GitOrigin-RevId: c62ef381d53593c61dbc0a6b9cfad058077601b8