[BreakoutBox] Don't use audio task runner for generator sinks in workers
This change avoids a thread jump when a MediaStreamTrackGenerator audio
sink runs on a worker.
One of the reasons for the thread jump is to prevent audio delivery
further down the pipeline from interfering with JS and possibly causing
jank. Since JS running on workers doesn't have access to the DOM, audio
delivery in the worker thread will not cause any user-perceived jank and
will save the overhead of the thread jump, which will result in better
performance, including user-perceived performance.

Bug: 1435005

Change-Id: I205b13c90904deae4c51c1b4ddb80df7531679ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4474631
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139498}
NOKEYCHECK=True
GitOrigin-RevId: 419c89e84daf6bcfaef274179cf04c27e151f1cc