VTonNav: Serialize device scale factor
When moving to the new renderer, we create the old pseudo elements
before we have a chance to initialize the device scale factor.

However, even if we do and it's a different value, we should know
what the previous renderer's device scale factor was. This patch
implements that.

R=khushalsagar@chromium.org

Fixed: 1434932
Change-Id: Ic7aa509b80ceadf9beedaf29c9a6e11bfa9fb38b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4497387
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139585}
NOKEYCHECK=True
GitOrigin-RevId: bc83e36a1698218e2e8136ab7943ee4eb8b10246