Build Attribution Reporting OS support on all platforms
Previously, build flags were used to restrict this to Android, but that
entailed significant complexity by infecting a lot of related code with
the need to itself be buildflag-protected. This will make it easier to
add OS support to other platforms in the future, if needed, and
increases test coverage.

This change introduces a no-op OS-level manager for platforms on which
OS support is disabled and removes layering violations between the
AttributionManager and AttributionOsLevelManager.

Change-Id: I7a984d6c77fd0312ad104c248d4278bb7666d99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4480909
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139825}
NOKEYCHECK=True
GitOrigin-RevId: a45d270b10caab11d3a8eec668eb53c8f83b5798