Pre-paint: Add BoxFragment() getter.
Use this one instead of having separate code paths for pre_paint_info_
vs. no pre_paint_info_.

If there's no pre_paint_info_, we don't have to use the LayoutBox as
fallback. Just use the first (and only) fragment instead.

As of CL:4476730 it is safe to assume that we won't enter the paint
property builder if we have a LayoutBox that lacks fragments.

Change-Id: Iaf5bbf802aae81d5e145cbdbe831896e9b1b9780
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4475148
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139979}
NOKEYCHECK=True
GitOrigin-RevId: 2163dc0e50465063035042dda9b9cba621253cd4