Prevent unnecessary throttling of 2D canvas rendering
The rate limiting mechanism for 2D canvas aims to ensure
that the compositor is never more than two frame behind the frame
currently being rendered by blink. This is meant to only apply
to animations that are not driven by "BeginFrame". For example
canvas animations driven by setTimeout, pointer events, etc.
It was recently noticed that some requestAnimationFrame driven
animations were triggering the rate limiter, which is not supposed
to happen.

The cause of the problem was that all canvas flushing operations
were being counted as frame boundaries. The MotionMark Images
benchmark, which calls putImageData on a large number of canvases
causes canvas paint ops to be flushed every time putImageData
is called. This caused the rate limiter to believe the GPU process
was falling behind, thus causing all canvases on the page to
start issuing sync queries in order to throttle the main thread's
execution. The sync query overhead was severely hurting performance.
With this CL, only PaintOp flushes for the purpose of pushing a
frame to the compositor count as a frame boundary.

This CL significantly increases Chrome's score on the MotionMark
Images test.

Bug: 1435066
Change-Id: I5a70a750c750a1efa3077a8dacbfb1e18816974d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4504839
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1140251}
NOKEYCHECK=True
GitOrigin-RevId: c5a4c1bdb564cbfda0eca2a4768e15672e9f1c7a