Change text highlight background color (attempt #2)
*** SHERIFFS: please don't revert this CL. If you find a flaky test
or two that looks related, please disable just that flaky test, and
then feel free to assign the bug to me. It's hard to land CLs with 250
baselines. Thanks. ***

See [1] for an initial attempt at this, which failed in some cases
such as... Windows. See [2]. This CL changes a few more default
locations, which turn out to be the ones that matter for AURA. This
CL should change the background color for <select> among other things
from #1E90FF to #1967D2.

In the process of making this change, and due to good review comments,
another bug was found, in that the wrong selection foreground color
was being used in the non-AURA code path. This CL also changes that
from pure black to pure white. Combined with the new background color,
that change improves selection contrast from 3.91 (black on blue) to
5.36 (white on blue). That triggers all new web test baselines.

I looked through all of the baselines in this CL, and all look good
to me. I flagged a few that seem to depend on the previously-black
foreground selection color - those might want to get updated to have
a different page background color. I've filed crbug.com/1443460 for
that, and I'll take care of it as a followup. Otherwise, I think the
rebaselines are clean.

----

Background: the renderer_preferences get used for AURA, if
use_custom_colors is true. That happens here:

https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:third_party/blink/renderer/core/exported/web_view_impl.cc;l=3394;drc=e39fffa6900a679961f5992b8f24a084853b811a

And the default renderer preferences has use_custom_colors true, and
sets the colors from here:

https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:third_party/blink/public/common/renderer_preferences/renderer_preferences.h;l=35;drc=e39fffa6900a679961f5992b8f24a084853b811a

or here:

https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/public/mojom/renderer_preferences.mojom;l=48;drc=047c7dc4ee1ce908d7fea38ca063fa2f80f92c77

This CL changes both of those to the new color.

It is unfortunate, but we don't use this code path for web_tests,
which is why most of the [1] CL changed but didn't actually have a
real effect on the browser. It turns out, the form controls browser
tests are the ones that should have changed, and they do here.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4288571
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=1343810#c37

Fixed: 1343810
Bug: 1443460
Change-Id: Ib1e81ed228780981008713dd907db9b18b2549e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4500340
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1140914}
NOKEYCHECK=True
GitOrigin-RevId: e3971a5b3be2c2a13fc83a8a91bb386a1da772f7