Remove expired Compositing.Renderer.FirstSurfaceActivationUpdateDuration
This metric has expired, and we gathered the desired information.

We are not going to pursue changes in this area. So removing the mertic
and the piping that communicated it from the Renderer to the Browser

OBSOLETE_HISTOGRAM[Compositing.Renderer.FirstSurfaceActivationUpdateDuration.{Surface}]=removing expired metric

Bug: 1426080
Change-Id: Ief87d47738c7d924323f78bf301799415c8ec3c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4512677
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141109}
NOKEYCHECK=True
GitOrigin-RevId: e90b4c68dfbd066edf8252a32e5ac85ac0d7dceb