Make WebRtcStatsReportIdl enabled-by-default.
At 50% Stable this is ready to launch. Making the feature
enabled-by-default in order to have a max_version for the GCL file. In
a follow-up CL I will delete the flag and any code path that would have
happened if the flag was disabled.

Bug: chromium:1414363
Change-Id: I97d7df50d5c1f2fe1bf3ceffbfc14c6c8eeefc73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4513943
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141313}
NOKEYCHECK=True
GitOrigin-RevId: f1fdde254c4d788cef24d34adf83f5d3e4222b3a