Support multiple URLs in Attribution Reporting OS registration
https://github.com/WICG/attribution-reporting-api/pull/780

To reduce the scope of this change, we handle the list of URLs
iteratively, rather than passing the list around as one unit. We can
investigate changing this in a followup CL in order to decrease the
number of Mojo and JNI calls.

Bug: 1442901
Change-Id: I7b942eec139a38c29f6e8539ffb5a332d75a4eae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4507802
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141322}
NOKEYCHECK=True
GitOrigin-RevId: 79fc516d04707df70055eb5dff54523622aed0b8