speculation rules: Force a style update on rule set removal.
With this change, we always wait for a microtask to update speculation
candidates. At that time, we either wait for style to be clean or if, as
with rule set removal, we want to ensure that we see the changes at a
predictable time, force it to be clean.

This case should be relatively rare and the system should otherwise
perform normally.

Bug: 1427644
Change-Id: I5af0377b6c861dbfb36955bcfa3a754089cab528
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4501939
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141397}
NOKEYCHECK=True
GitOrigin-RevId: df68e37d623400f4a91d8cf4827f3c5c90dcea65