[webrtc-internals] Remove asterisk indicating non-standard stats.
As much as I like this asterisk to indicate that a metric visible in
webrtc-internals is NOT exposed to JavaScript, this only applies to a
handful of metrics and this CL unblocks some WebRTC cleanup:
https://crbug.com/webrtc/15162

The is_standardized() and StatExposureCriteria are no longer needed
thanks to WebRtcStatsReportIdl having launched, so there is no longer a
need to have multiple RTCStatsMember classes with inheritence.

This CL removes the last reference to
RTCNonStandardStatsMember-specific code paths.

Bug: webrtc:15162
Change-Id: I5c80a8b31e0755f3fcd8454dfb6316d37c1fb55c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4516539
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141423}
NOKEYCHECK=True
GitOrigin-RevId: e0d9f50123d0f0b81bc19deacfa9bab7c299e8af