Increase canvas resource recycling capacity
CanvasResourceProvider is occasionally starved of resources during
normal operation due to a race between commit completion and the
beginning of the next frame. When the task order is inverted, the
resource provider may run out of cached resources, causing it to
synchronously create new SharedImages.

The stochasticity of the race was causing the MotionMark Images score
to be particularly noisy. This CL implements a short term solution
which is to increase the max cached resources to 3 (from 2).  Longer
term, it would be best to implement a resource manager that can
increase and reduce cache capacity adaptively and pool resources
between different canvases in order to degrade performance more
gracefully when capacity is exceeded.

Bug: 1435066
Change-Id: I8e814e3d62349715f923e2985d40cf9424ff3dea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4519163
Auto-Submit: Justin Novosad <junov@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141657}
NOKEYCHECK=True
GitOrigin-RevId: 692d013807419afdbb1bb5ac39ad73cbaa3ae96d