Remove LazyImageHelper::Eligibility and simplify code
LazyImageHelper::DetermineEligibilityAndTrackVisibilityMetrics had
leftover support for multiple types of eligibility from when there was
both developer as well as automatic lazy loading. This patch removes
LazyImageHelper::Eligibility and splits the eligibility function into
ShouldDeferImageLoad and StartMonitoringVisibilityMetrics. In a
followup patch, LazyImageHelper will be removed.

There is a minor metrics change with this patch when the lazy load
setting is disabled, which is rare. Previously, we would record
Blink.VisibleBeforeLoaded.LazyLoadImages.{Above, Below}TheFold2
metrics when the lazy loading setting was disabled, but now we do not.
The new behavior matches the metric description.

Bug: 1441062
Change-Id: I62ed3dd8e7f3d25783a940e97cc6c8547223505c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518845
Reviewed-by: Traian Captan <tcaptan@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141689}
NOKEYCHECK=True
GitOrigin-RevId: 5861b03d661869c99c1ea3289cb3a63bf38eb228