AbortSignal: Allow internal signals as source signals
NavigateEvent's internally constructed AbortSignal is exposed to
userland, which means it can be passed to AbortSignal.any(). This isn't
currently supported (it'll DCHECK), but is generally safe as long as
the signal isn't aborted in userland or made to follow other signals
later, which is the case. This will be made safer in follow-up CLs
when we make SignalAbort private and replace Follow with the composite
signal constructor.

Bug: 1323391
Change-Id: I2a4b477bdb7026be9e0b28d733c100614bb34d6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518012
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1141729}
NOKEYCHECK=True
GitOrigin-RevId: a0cdf065ea3faf831add2f28a253860db2b8970c